+++ Tp Final Hernandez Martin Lorenzo Leandro.depend
@@ -114,10 +114,10 @@
 1604584682 c:\users\llorenzo\desktop\git tp final\archivoclientes.h
 	"archivoConsumos.h"
 
-1604584682 c:\users\llorenzo\desktop\git tp final\archivoconsumos.h
+1604608812 c:\users\llorenzo\desktop\git tp final\archivoconsumos.h
 	"archivoClientes.h"
 
-1604607569 c:\users\llorenzo\desktop\git tp final\lista.h
+1604609310 c:\users\llorenzo\desktop\git tp final\lista.h
 	"archivoConsumos.h"
 	"archivoClientes.h"
 
@@ -129,7 +129,7 @@
 	"archivoConsumos.h"
 	<stdio.h>
 
-1604584682 source:c:\users\llorenzo\desktop\git tp final\main.c
+1604609310 source:c:\users\llorenzo\desktop\git tp final\main.c
 	<stdio.h>
 	<stdlib.h>
 	<string.h>
@@ -139,8 +139,16 @@
 	"archivoClientes.h"
 	"archivoConsumos.h"
 	"lista.h"
+	"utilidades.h"
 
-1604607569 source:c:\users\llorenzo\desktop\git tp final\lista.c
+1604609346 source:c:\users\llorenzo\desktop\git tp final\lista.c
 	"lista.h"
 	<stdio.h>
+
+1604608882 source:c:\users\llorenzo\desktop\git tp final\utilidades.c
+	"utilidades.h"
+
+1604608827 c:\users\llorenzo\desktop\git tp final\utilidades.h
+	"archivoConsumos.h"
+	"archivoClientes.h"
